Moved Pearson distance to correlations.rs
