+++ crates/distances/src/vectors/mod.rs
@@ -7,8 +7,10 @@
 //! shorter vector will be ignored.
 
 mod angular;
+mod correlations;
 mod lp_norms;
 pub(crate) mod utils;
 
 pub use angular::{bray_curtis, canberra, cosine, hamming};
-pub use lp_norms::{chebyshev, euclidean, euclidean_sq, l3_norm, l4_norm, manhattan, minkowski, minkowski_p, pearson};
+pub use correlations::pearson;
+pub use lp_norms::{chebyshev, euclidean, euclidean_sq, l3_norm, l4_norm, manhattan, minkowski, minkowski_p};
